change position type from integer to long

+++ unicorecmsebola/tests/test_views.py
@@ -33,7 +33,7 @@ class TestViews(UnicoreTestCase):
                     'index': 'not_analyzed',
                 },
                 'position': {
-                    'type': 'integer'
+                    'type': 'long'
                 }
             }
         })
